django-backend/requirements.txt
appengine-python-standard~=1.0.0
Django~=3.2.13
django-environ~=0.9.0
django-ninja~=0.18.0
email-validator~=1.2.1
firebase-admin~=5.2.0
google-cloud-firestore~=2.5.3
google-cloud-secret-manager~=2.11.1
mailerlite-api-python~=0.10.0
psycopg2~=2.9.3